Remove `half_node` and `nest_cfgs` from hugr crate.
Move one test into the hugr-passes crate, and refactor `depth()` into
`utils`, to avoid code duplication.
