+++ hugr/src/algorithm.rs
@@ -1,5 +1,3 @@
 //! Algorithms using the Hugr.
 
 pub mod const_fold;
-mod half_node;
-pub mod nest_cfgs;
